% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2.R
\docType{data}
\name{allfh}
\alias{allfh}
\title{Diet and Condition Data}
\format{
A data frame with 1423676 rows and 90 variables:
\describe{
\item{\code{pynam}}{character Prey scientific name}
\item{\code{gencat}}{character Prey general category (shortened)}
\item{\code{gensci}}{character Prey general category (full)}
\item{\code{analcat}}{character Prey analytical category (shortened)}
\item{\code{analsci}}{character Prey analytical category (full)}
\item{\code{pdlen}}{double Predtaor length}
\item{\code{pdwgt}}{integer Predator weight}
\item{\code{stratum}}{integer Stratum}
\item{\code{surftemp}}{double Surface water temperature}
\item{\code{bottemp}}{double Bottom water temperature}
\item{\code{year}}{integer Year}
\item{\code{season}}{integer Season}
\item{\code{pyamtw}}{double Prey weight}
\item{\code{Species}}{character Predator species}
\item{\code{Region}}{character Predator region}
\item{\code{fish_id}}{character Predator ID number}

\item{\code{obs}}{integer COLUMN_DESCRIPTION}
\item{\code{pyspp}}{double COLUMN_DESCRIPTION}
\item{\code{pyabbr}}{integer COLUMN_DESCRIPTION}
\item{\code{modcat}}{integer COLUMN_DESCRIPTION}
\item{\code{collcat}}{integer COLUMN_DESCRIPTION}
\item{\code{collsci}}{integer COLUMN_DESCRIPTION}
\item{\code{station}}{integer COLUMN_DESCRIPTION}
\item{\code{pdsex}}{integer COLUMN_DESCRIPTION}
\item{\code{pdid}}{integer COLUMN_DESCRIPTION}
\item{\code{fhmat}}{integer COLUMN_DESCRIPTION}
\item{\code{pdmat}}{integer COLUMN_DESCRIPTION}
\item{\code{pdsvol}}{double COLUMN_DESCRIPTION}
\item{\code{pdswgt}}{double COLUMN_DESCRIPTION}
\item{\code{fhdat}}{integer COLUMN_DESCRIPTION}
\item{\code{cruise}}{integer COLUMN_DESCRIPTION}
\item{\code{cruise6}}{integer COLUMN_DESCRIPTION}
\item{\code{svspp}}{integer COLUMN_DESCRIPTION}
\item{\code{catsex}}{integer COLUMN_DESCRIPTION}
\item{\code{cf}}{double COLUMN_DESCRIPTION}
\item{\code{sizecat}}{integer COLUMN_DESCRIPTION}
\item{\code{len10cat}}{integer COLUMN_DESCRIPTION}
\item{\code{len5cat}}{integer COLUMN_DESCRIPTION}
\item{\code{pdgutw}}{double COLUMN_DESCRIPTION}
\item{\code{pdgutv}}{double COLUMN_DESCRIPTION}
\item{\code{ttime}}{integer COLUMN_DESCRIPTION}
\item{\code{towdur}}{double COLUMN_DESCRIPTION}
\item{\code{setdepth}}{integer COLUMN_DESCRIPTION}
\item{\code{airtemp}}{double COLUMN_DESCRIPTION}
\item{\code{svgear}}{integer COLUMN_DESCRIPTION}
\item{\code{tow}}{integer COLUMN_DESCRIPTION}
\item{\code{beglat}}{double COLUMN_DESCRIPTION}
\item{\code{beglon}}{double COLUMN_DESCRIPTION}
\item{\code{hour}}{integer COLUMN_DESCRIPTION}
\item{\code{minute}}{integer COLUMN_DESCRIPTION}
\item{\code{month}}{integer COLUMN_DESCRIPTION}
\item{\code{day}}{integer COLUMN_DESCRIPTION}
\item{\code{haul}}{integer COLUMN_DESCRIPTION}
\item{\code{gearcond}}{integer COLUMN_DESCRIPTION}
\item{\code{area}}{integer COLUMN_DESCRIPTION}
\item{\code{statype}}{integer COLUMN_DESCRIPTION}
\item{\code{towtime}}{integer COLUMN_DESCRIPTION}
\item{\code{purcode}}{integer COLUMN_DESCRIPTION}
\item{\code{status_code}}{integer COLUMN_DESCRIPTION}
\item{\code{yr2block}}{double COLUMN_DESCRIPTION}
\item{\code{yr3block}}{integer COLUMN_DESCRIPTION}
\item{\code{yr5block}}{integer COLUMN_DESCRIPTION}
\item{\code{decade}}{integer COLUMN_DESCRIPTION}
\item{\code{garea}}{integer COLUMN_DESCRIPTION}
\item{\code{geoarea}}{integer COLUMN_DESCRIPTION}
\item{\code{declat}}{double COLUMN_DESCRIPTION}
\item{\code{declon}}{double COLUMN_DESCRIPTION}
\item{\code{catwgt}}{double COLUMN_DESCRIPTION}
\item{\code{catnum}}{integer COLUMN_DESCRIPTION}
\item{\code{numlen}}{integer COLUMN_DESCRIPTION}
\item{\code{pynum}}{integer COLUMN_DESCRIPTION}
\item{\code{pyperi}}{integer COLUMN_DESCRIPTION}
\item{\code{pywgti}}{double COLUMN_DESCRIPTION}
\item{\code{pyvoli}}{double COLUMN_DESCRIPTION}
\item{\code{pylen}}{logical COLUMN_DESCRIPTION}
\item{\code{pyamtv}}{double COLUMN_DESCRIPTION}
\item{\code{perpyv}}{double COLUMN_DESCRIPTION}
\item{\code{perpyw}}{double COLUMN_DESCRIPTION}
\item{\code{pdscinam}}{integer COLUMN_DESCRIPTION}
\item{\code{pdcomnam}}{integer COLUMN_DESCRIPTION}
\item{\code{tot_tows_spp_stratum}}{integer COLUMN_DESCRIPTION}
\item{\code{tot_catnum_stratum}}{integer COLUMN_DESCRIPTION}
\item{\code{tot_catwgt_stratum}}{double COLUMN_DESCRIPTION}
\item{\code{stratum_area}}{integer COLUMN_DESCRIPTION}
\item{\code{pycomnam2}}{integer COLUMN_DESCRIPTION}
\item{\code{gencom2}}{integer COLUMN_DESCRIPTION}
\item{\code{analcom3}}{integer COLUMN_DESCRIPTION}
\item{\code{collcom3}}{logical COLUMN_DESCRIPTION}
\item{\code{spst}}{character COLUMN_DESCRIPTION}
\item{\code{stock_area}}{character COLUMN_DESCRIPTION}
}
}
\source{
\url{https://github.com/Laurels1/Condition/blob/master/data/allfh.RData}
}
\usage{
allfh
}
\description{
This data was derived from NEFSC bottom trawl data. It contains stomach content information and length-weight information.
}
\details{
DETAILS
}
\keyword{datasets}
