% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2.R
\docType{data}
\name{allfh_small}
\alias{allfh_small}
\title{Diet and Condition Data}
\format{
A data frame with 1423676 rows and 16 variables:
\describe{
\item{\code{pynam}}{character Prey scientific name}
\item{\code{gencat}}{character Prey general category (shortened)}
\item{\code{gensci}}{character Prey general category (full)}
\item{\code{analcat}}{character Prey analytical category (shortened)}
\item{\code{analsci}}{character Prey analytical category (full)}
\item{\code{pdlen}}{double Predtaor length}
\item{\code{pdwgt}}{integer Predator weight}
\item{\code{stratum}}{integer Stratum}
\item{\code{surftemp}}{double Surface water temperature}
\item{\code{bottemp}}{double Bottom water temperature}
\item{\code{year}}{integer Year}
\item{\code{season}}{integer Season}
\item{\code{pyamtw}}{double Prey weight}
\item{\code{Species}}{character Predator species}
\item{\code{Region}}{character Predator region}
\item{\code{fish_id}}{character Predator ID number}
}
}
\source{
\url{https://github.com/Laurels1/Condition/blob/master/data/allfh.RData}
}
\usage{
allfh_small
}
\description{
This data was derived from NEFSC bottom trawl data. It contains stomach content information and length-weight information.
}
\details{
DETAILS
}
\keyword{datasets}
