% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{asmt}
\alias{asmt}
\title{Stock Assessment Data}
\format{
A data frame with 35636 rows and 10 variables:
\describe{
\item{\code{Species}}{character The species' common name}
\item{\code{Region}}{character The stock region}
\item{\code{Year}}{double The year of the value estimate}
\item{\code{Value}}{double The value of the metric}
\item{\code{Metric}}{character The stock assessment metric (catch, F, abundance, or recruitment)}
\item{\code{Description}}{character Details about the metric}
\item{\code{Units}}{character The units of the metric}
\item{\code{AssessmentYear}}{double The year that the assessment was conducted}
\item{\code{Age}}{character The age of the fish (recruitment/abundance only)}
\item{\code{Category}}{character COLUMN_DESCRIPTION}
}
}
\source{
\url{https://github.com/NOAA-EDAB/assessmentdata}
}
\usage{
asmt
}
\description{
This data set contains stock assessment data on several US stocks.
}
\details{
DETAILS
}
\keyword{datasets}
