% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2.R
\docType{data}
\name{asmt_sum}
\alias{asmt_sum}
\title{Stock Assessment Summary Data}
\format{
A data frame with 337 rows and 61 variables:
\describe{

\item{\code{FSSI Stock?}}{character Is the stock included in the Fish Stock Sustainability Index?}
\item{\code{Assessment Year}}{integer Assessment year}
\item{\code{Last Data Year}}{integer Last data year used in assessment}
\item{\code{Review Result}}{character Results of assessment review by the Council}
\item{\code{Life History Data}}{integer Biological data rating (before 2019): 0 = no data ; 4 = sufficient data for informed modeling}
\item{\code{Biological Input Data}}{logical Biological data rating (2019 onwards): 0 = no data ; 5 = detailed data with no gaps}
\item{\code{Ecosystem Linkage}}{integer Use of ecosystem linkage data in stock assessment model: 0 = not used ; 5 = model is linked to ecosystem processes}
\item{\code{Composition Input Data}}{integer Availability of size/age data for the stock assessment model: 0 = none ; 5 = complete data with spatial resolution}
\item{\code{F/Fmsy}}{double Fishing pressure relative to the fishing pressure for maximum sustainable yield}
\item{\code{B/Bmsy}}{double Biomass relative to the biomass for maximum sustainable yield}
\item{\code{Species}}{character Species common name}
\item{\code{Region}}{character Stock region}

\item{\code{Stock Name}}{character COLUMN_DESCRIPTION}
\item{\code{Jurisdiction}}{character COLUMN_DESCRIPTION}
\item{\code{FMP}}{character COLUMN_DESCRIPTION}
\item{\code{Science Center}}{character COLUMN_DESCRIPTION}
\item{\code{Regional Ecosystem}}{character COLUMN_DESCRIPTION}
\item{\code{ITIS Taxon Serial Number}}{integer COLUMN_DESCRIPTION}
\item{\code{Scientific Name}}{character COLUMN_DESCRIPTION}
\item{\code{Common Name}}{character COLUMN_DESCRIPTION}
\item{\code{Stock Area}}{character COLUMN_DESCRIPTION}
\item{\code{Assessment Month}}{integer COLUMN_DESCRIPTION}
\item{\code{Update Type}}{character COLUMN_DESCRIPTION}
\item{\code{Assessment Model}}{character COLUMN_DESCRIPTION}
\item{\code{Model Version}}{character COLUMN_DESCRIPTION}
\item{\code{Lead Lab}}{character COLUMN_DESCRIPTION}
\item{\code{Citation}}{character COLUMN_DESCRIPTION}
\item{\code{Final Assessment Report 1}}{character COLUMN_DESCRIPTION}
\item{\code{Final Assessment Report 2}}{character COLUMN_DESCRIPTION}
\item{\code{Point of Contact}}{character COLUMN_DESCRIPTION}
\item{\code{Abundance Data}}{integer COLUMN_DESCRIPTION}
\item{\code{Catch Data}}{integer COLUMN_DESCRIPTION}
\item{\code{Assessment Level}}{integer COLUMN_DESCRIPTION}
\item{\code{Assessment Frequency}}{integer COLUMN_DESCRIPTION}
\item{\code{Assessment Type}}{character COLUMN_DESCRIPTION}
\item{\code{Model Category}}{integer COLUMN_DESCRIPTION}
\item{\code{Catch Input Data}}{integer COLUMN_DESCRIPTION}
\item{\code{Abundance Input Data}}{integer COLUMN_DESCRIPTION}
\item{\code{F Year}}{integer COLUMN_DESCRIPTION}
\item{\code{Estimated F}}{double COLUMN_DESCRIPTION}
\item{\code{F Unit}}{character COLUMN_DESCRIPTION}
\item{\code{F Basis}}{character COLUMN_DESCRIPTION}
\item{\code{Flimit}}{double COLUMN_DESCRIPTION}
\item{\code{Flimit Basis}}{character COLUMN_DESCRIPTION}
\item{\code{Fmsy}}{double COLUMN_DESCRIPTION}
\item{\code{Fmsy Basis}}{character COLUMN_DESCRIPTION}
\item{\code{F/Flimit}}{double COLUMN_DESCRIPTION}
\item{\code{Ftarget}}{double COLUMN_DESCRIPTION}
\item{\code{Ftarget Basis}}{character COLUMN_DESCRIPTION}
\item{\code{F/Ftarget}}{double COLUMN_DESCRIPTION}
\item{\code{B Year}}{integer COLUMN_DESCRIPTION}
\item{\code{Estimated B}}{double COLUMN_DESCRIPTION}
\item{\code{B Unit}}{character COLUMN_DESCRIPTION}
\item{\code{B Basis}}{character COLUMN_DESCRIPTION}
\item{\code{Blimit}}{double COLUMN_DESCRIPTION}
\item{\code{Blimit Basis}}{character COLUMN_DESCRIPTION}
\item{\code{Bmsy}}{double COLUMN_DESCRIPTION}
\item{\code{Bmsy Basis}}{character COLUMN_DESCRIPTION}
\item{\code{B/Blimit}}{double COLUMN_DESCRIPTION}
\item{\code{MSY}}{double COLUMN_DESCRIPTION}
\item{\code{MSY Unit}}{character COLUMN_DESCRIPTION}
}
}
\source{
\url{https://github.com/NOAA-EDAB/assessmentdata}
}
\usage{
asmt_sum
}
\description{
This data set contains summary stock assessment metrics on several US stocks.
}
\details{
For more details see \url{https://www.st.nmfs.noaa.gov/stocksmart/StockSMART_DataDictionary.pdf}
}
\keyword{datasets}
