% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bio_survey}
\alias{bio_survey}
\title{NEFSC Bottom Trawl Data with Bio Data}
\format{
A data frame with 1103575 rows and 28 variables:
\describe{
\item{\code{CRUISE6}}{character Cruise ID}
\item{\code{STATION}}{character Station ID}
\item{\code{STRATUM}}{character Stratum}
\item{\code{SVSPP}}{character Northeast Resource Surveys Species code}
\item{\code{CATCHSEX}}{character Species-specific numeric sex code}
\item{\code{LENGTH}}{double Length}
\item{\code{TOW}}{character Tow ID}
\item{\code{SVVESSEL}}{character Vessel ID}
\item{\code{YEAR}}{double Year}
\item{\code{SEASON}}{character Season}
\item{\code{LAT}}{double Latitude}
\item{\code{LON}}{double Longitude}
\item{\code{EST_TOWDATE}}{double Date with estimated time}
\item{\code{DEPTH}}{double Depth}
\item{\code{SURFTEMP}}{double Surface water temperature}
\item{\code{SURFSALIN}}{double Surface salinity}
\item{\code{BOTTEMP}}{double Bottom water temperature}
\item{\code{BOTSALIN}}{double Bottom salinity}
\item{\code{ABUNDANCE}}{double Abundance (number of fish caught on a single tow; redundant for rows with the same CRUISE6, STATION, and TOW)}
\item{\code{BIOMASS}}{double Biomass (biomass caught on a single tow; redundant for rows with the same SVSPP, CRUISE6, STATION, and TOW)}
\item{\code{NUMLEN}}{double Number of fish at the specified length (redundant for rows with the same SVSPP, CRUISE6, STATION, TOW, and LENGTH)}
\item{\code{INDID}}{double Individual fish ID; unknown if actually unique}
\item{\code{INDWT}}{double Individual fish weight}
\item{\code{SEX}}{character Sex}
\item{\code{MATURITY}}{character Maturity}
\item{\code{AGE}}{double Age}
\item{\code{STOM_VOLUME}}{double Stomach volume}
\item{\code{STOM_WGT}}{double Stomach weight}
}
}
\usage{
bio_survey
}
\description{
This data is NEFSC bottom trawl data with bio data.
}
\details{
\itemize{
\item Function call: survdat::get_survdat_data(channel = channel, filterByYear = NA, all.season = TRUE, shg.check = T, conversion.factor = T, use.SAD = F, getBio = TRUE, getLengths = T)
\item Pull date: Tue Mar 02 16:12:40 2021
}
}
\keyword{datasets}
