% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{character_to_factor}
\alias{character_to_factor}
\title{Changes the class of \code{character} columns to \code{factor}}
\usage{
character_to_factor(x)
}
\arguments{
\item{x}{A data frame or tibble}
}
\value{
A data frame
}
\description{
This function changes \code{character} columns to class \code{factor} for better rendering as an html table with the \code{DT} package.
}
