% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{climate_risk}
\alias{climate_risk}
\title{Overall Climate Risk}
\format{
A data frame with 82 rows and 3 variables:
\describe{
\item{\code{Species}}{character Species' common name}
\item{\code{Overall_climate_vulnerability}}{character Overall climate vulnerability score}
\item{\code{Certainty}}{character Certainty score}
}
}
\source{
\url{https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0146756}
}
\usage{
climate_risk
}
\description{
This is the overall climate risk score data from Hare et al. 2016.
}
\details{
DETAILS
}
\keyword{datasets}
