% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{climate_vulnerability}
\alias{climate_vulnerability}
\title{Climate Vulnerability}
\format{
A data frame with 1968 rows and 8 variables:
\describe{
\item{\code{Species}}{character Species' common name}
\item{\code{Functional.Group}}{character Species guild}
\item{\code{Attribute}}{character COLUMN_DESCRIPTION}
\item{\code{Attribute.Category}}{character COLUMN_DESCRIPTION}
\item{\code{Low}}{integer Number of experts ranking attribute risk as low}
\item{\code{Moderate}}{integer Number of experts ranking attribute risk as moderate}
\item{\code{High}}{integer Number of experts ranking attribute risk as high}
\item{\code{Very.High}}{integer Number of experts ranking attribute risk as very high}
}
}
\source{
\url{https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0146756}
}
\usage{
climate_vulnerability
}
\description{
This is the climate vulnerability data from Hare et al. 2016.
}
\details{
DETAILS
}
\keyword{datasets}
