% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{com_catch}
\alias{com_catch}
\title{Commercial Catch and Revenue Data}
\format{
A data frame with 31637 rows and 9 variables:
\describe{
\item{\code{Species}}{character Species}
\item{\code{Year}}{integer Year}
\item{\code{State}}{character State}
\item{\code{Pounds}}{double Pounds landed}
\item{\code{Dollars}}{double Dollars}
\item{\code{Dollars_adj}}{double Dollars (2019 value)}
}
}
\usage{
com_catch
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
