% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2.R
\docType{data}
\name{com_gear}
\alias{com_gear}
\title{DATASET_TITLE}
\format{
A data frame with 277963 rows and 10 variables:
\describe{
\item{\code{YEAR}}{integer COLUMN_DESCRIPTION}
\item{\code{NESPP3}}{double COLUMN_DESCRIPTION}
\item{\code{QY}}{double COLUMN_DESCRIPTION}
\item{\code{GEAR}}{integer COLUMN_DESCRIPTION}
\item{\code{SIZE}}{integer COLUMN_DESCRIPTION}
\item{\code{EPU}}{integer COLUMN_DESCRIPTION}
\item{\code{UTILCD}}{double COLUMN_DESCRIPTION}
\item{\code{SPPLIVMT}}{double COLUMN_DESCRIPTION}
\item{\code{SPPVALUE}}{double COLUMN_DESCRIPTION}
\item{\code{US}}{logical COLUMN_DESCRIPTION}
}
}
\usage{
com_gear
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
