% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{com_palette}
\alias{com_palette}
\title{Commercial landings data color palette}
\format{
A data frame with 11 rows and 2 variables:
\describe{
\item{\code{state_id}}{character State}
\item{\code{color}}{character Color (hex code)}
}
}
\usage{
com_palette
}
\description{
A color palette to use with commercial data
}
\details{
Created from \code{nmfspalette} package
}
\keyword{datasets}
