% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survdat_functions.R
\name{common_names_survdat}
\alias{common_names_survdat}
\title{Format \code{survdat} to add common names}
\usage{
common_names_survdat(survdat_pull_type = "all")
}
\arguments{
\item{survdat_pull_type}{A desired dataset either all records or a set containing only records with biological information.}
}
\value{
A tibble
}
\description{
This function appends \code{survdat} with common names for easier manipulation. By default returns all survdata that is not constrained by biological information. If provided "bio" will return a limited dataset containing only records with biological information such as age and sex.
}
