% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cond}
\alias{cond}
\title{Condition data}
\format{
A data frame with 3595 rows and 7 variables:
\describe{
\item{\code{Species}}{character Species}
\item{\code{EPU}}{character Environmental Protection Unit}
\item{\code{sex}}{character Sex}
\item{\code{YEAR}}{integer Year}
\item{\code{MeanCond}}{double Mean condition}
\item{\code{nCond}}{integer COLUMN_DESCRIPTION}
\item{\code{n}}{integer COLUMN_DESCRIPTION}
}
}
\source{
2018 data from: \url{https://github.com/Laurels1/Condition}
}
\usage{
cond
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
