% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_functions.R
\name{correlation_summary}
\alias{correlation_summary}
\title{Produces a summary table of correlations}
\usage{
correlation_summary(data, lag = 0)
}
\arguments{
\item{data}{The output of \code{data_prep()}.}

\item{lag}{The number of years to lag the correlation by. Defaults to 0.}
}
\value{
A tibble
}
\description{
This function creates summary table of a correlation between stock data and \code{ecodata} data. Suggest to use with multiple \code{ecodata} indicators, and create a master data set by appending results.
}
