% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ecodata_indicators.R
\name{create_coldpool_extent}
\alias{create_coldpool_extent}
\title{Create Cold Pool Extent}
\usage{
create_coldpool_extent(data)
}
\arguments{
\item{data}{Filepath to the indicator data}

\item{return}{Boolean. Whether to return the indicator as an object in the global environment}
}
\value{
Saves R object \code{coldpool_extent}, returns cold pool extent indicator
}
\description{
This function generates a cold pool extent indicator from ecodata R package to an R object
Cold pool extent:  total area where bottom temperatures remain below 10 °C for at least 2 months between June and September
Positive values == Larger
Negative values == Smaller
}
