% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ecodata_indicators.R
\name{create_coldpool_index}
\alias{create_coldpool_index}
\title{Create Cold Pool Index}
\usage{
create_coldpool_index(data)
}
\arguments{
\item{data}{Filepath to the indicator data}

\item{return}{Boolean. Whether to return the indicator as an object in the global environment}
}
\value{
Saves R object \code{coldpool_index}, returns cold pool index indicator
}
\description{
This function generates a cold pool index indicator from ecodata R package to an R object
Cold pool index: quantifies the interannual strength of the cold pool
Positive values == colder
Negative values == warmer
}
