% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ecodata_indicators.R
\name{create_coldpool_persistence}
\alias{create_coldpool_persistence}
\title{Create Cold Pool Persistence}
\usage{
create_coldpool_persistence(data)
}
\arguments{
\item{data}{Filepath to the indicator data}

\item{return}{Boolean. Whether to return the indicator as an object in the global environment}
}
\value{
Saves R object \code{coldpool_persistence}, returns cold pool persistence indicator
}
\description{
This function generates a cold pool persistence indicator from ecodata R package to an R object
Cold pool persistence: duration of the cold pool, which ends when bottom temperature rises above 10 °C after it is formed each year
Positive values == Longer
Negative values == Shorter
}
