% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ecodata_indicators.R
\name{create_gsi}
\alias{create_gsi}
\title{Create Gulf stream Index}
\usage{
create_gsi(data)
}
\arguments{
\item{data}{Filepath to the indicator data}

\item{return}{Boolean. Whether to return the indicator as an object in the global environment}
}
\value{
Saves R object \code{gsi}, returns Gulf Stream Index indicator
}
\description{
This function generates a Gulf Stream index indicator from ecodata R package to an R object
Gulf Stream Index: measure of the Gulf Stream position relative to the mean position
Positive values == More northerly
Negative values == More southerly
}
