% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ecodata_indicators.R
\name{create_wcr}
\alias{create_wcr}
\title{Create Warm Core Rings}
\usage{
create_wcr(data)
}
\arguments{
\item{data}{Filepath to the indicator data}

\item{return}{Boolean. Whether to return the indicator as an object in the global environment}
}
\value{
Saves R object \code{wcr}, returns warm core rings indicator
}
\description{
This function generates a warm core rings indicator from ecodata R package to an R object
Warm Core Rings: number of warm core ring formations
}
