% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_functions.R
\name{data_prep}
\alias{data_prep}
\title{Prepares data for correlation analysis}
\usage{
data_prep(stock_data, eco_data, lag_data = 0)
}
\arguments{
\item{stock_data}{Data about a single stock (one species, one region) subsetted from \code{assessmentdata::stockAssessmentData}}

\item{eco_data}{A data table from \code{ecodata}. May require pre-processing to standardize format.}

\item{lag_data}{The number of years to lag the correlation by. Defaults to 0.}
}
\value{
A tibble
}
\description{
This function prepares stock data and \code{ecodata} data for correlation analysis and returns linear correlation p-values. A helper function for \code{plot_correlation}, \code{correlation_data}, and \code{correlation_summary}.
}
