% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survdat_functions.R
\name{data_summary_5yr}
\alias{data_summary_5yr}
\title{Summarise the past 5 years of \code{survdat} data}
\usage{
data_summary_5yr(x)
}
\arguments{
\item{x}{A data frame or tibble, containing data on one species. The output of \code{get_var_data}.}
}
\value{
A \code{DT::datatable}
}
\description{
This function summarises the past five years of \code{survdat} data, grouping by by region. Data must be pre-processed with \code{get_var_data}.
}
