% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_files}
\alias{find_files}
\title{Locate files containing a text string}
\usage{
find_files(text)
}
\arguments{
\item{text}{A text string to search for.}
}
\value{
A vector of names of files that contain the string.
}
\description{
This function locates all .R and .Rmd files that contain a text string.
}
