% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_numbers}
\alias{format_numbers}
\title{Change the class of a vector to \code{numeric}}
\usage{
format_numbers(x)
}
\arguments{
\item{x}{A vector}
}
\value{
A vector
}
\description{
This function changes a vector to class \code{int} or \code{dbl}. Used when numerics are not read in properly.
}
