% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survdat_functions.R
\name{generate_len_plot}
\alias{generate_len_plot}
\title{A wrapper function to plot \code{survdat} length data}
\usage{
generate_len_plot(x)
}
\arguments{
\item{x}{A \code{survdat} data frame or tibble, containing data on one species.}
}
\value{
A ggplot
}
\description{
This function is a wrapper for \code{get_len_data} and \code{plot_len}.
}
