% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survdat_functions.R
\name{generate_len_table}
\alias{generate_len_table}
\title{A wrapper function to summarise \code{survdat} length data}
\usage{
generate_len_table(x)
}
\arguments{
\item{x}{A \code{survdat} data frame or tibble, containing data on one species.}
}
\value{
A \code{knitr::kable}
}
\description{
This function is a wrapper for \code{get_len_data_tbl}, \code{len_tbl_data}, and \code{len_tbl_data_5yr}. Returns a summary table.
}
