% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survdat_functions.R
\name{generate_plot}
\alias{generate_plot}
\title{A wrapper function to plot \code{survdat} data}
\usage{
generate_plot(x, ytitle = "", variable)
}
\arguments{
\item{x}{A \code{survdat} data frame or tibble, containing data on one species.}

\item{ytitle}{The title of the y-axis. Defaults to "".}

\item{variable}{The \code{survdat} measurement of interest. Must be a column name of the \code{survdat} data set.}
}
\value{
A ggplot
}
\description{
This function is a wrapper for \code{get_var_data} and \code{plot_variable}.
}
