% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survdat_functions.R
\name{generate_table}
\alias{generate_table}
\title{A wrapper function to summarise \code{survdat} data}
\usage{
generate_table(x, variable, cap = "")
}
\arguments{
\item{x}{A \code{survdat} data frame or tibble, containing data on one species.}

\item{variable}{The \code{survdat} measurement of interest. Must be a column name of the \code{survdat} data set.}

\item{cap}{A caption for the table. Defaults to "".}
}
\value{
A \code{DT::datatable}
}
\description{
This function is a wrapper for \code{get_var_data}, \code{data_summary}, and \code{data_summary_5yr}.
}
