% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{get_diet_plot}
\alias{get_diet_plot}
\title{Plot diet composition}
\usage{
get_diet_plot(data)
}
\arguments{
\item{x}{A data frame or tibble, containing data on one species. Data from \code{allfh}.}
}
\value{
A ggplot
}
\description{
This function plots the proportional composition of a species' diet.
}
