% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{get_diet_table}
\alias{get_diet_table}
\title{Create a table of diet data}
\usage{
get_diet_table(data, type = "html")
}
\arguments{
\item{data}{A data frame or tibble, containing data on one species. Data from \code{allfh}.}

\item{type}{The file type of the output. One of c("html", "word")}
}
\value{
A \code{DT::datatable} or a \code{knitr::kable}
}
\description{
This function creates a table of the proportional composition of a species' diet.
}
