% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_functions.R
\name{get_latlong}
\alias{get_latlong}
\title{Create a summary table of species range}
\usage{
get_latlong(x, data, shapefile)
}
\arguments{
\item{x}{The common name of the species.}

\item{data}{A data frame or tibble, containing data on one species. Must contain season and strata information.}

\item{shapefile}{A shapefile to convert strata ID to a shape.}
}
\value{
A ggplot
}
\description{
This function creates a summary table of species distribution (latitude/longitude range).
}
