% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survdat_functions.R
\name{get_len_data}
\alias{get_len_data}
\title{Format \code{survdat} length data}
\usage{
get_len_data(x)
}
\arguments{
\item{x}{A \code{survdat} data frame or tibble, containing data on one ospecies.}
}
\value{
A tibble
}
\description{
This function formats \code{survdat} length data for subsequent plotting and analysis. Data from unique observations of a stock (species, region) are averaged by year. Returns lengths in a tidy format.
}
