% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survdat_functions.R
\name{get_len_data_risk}
\alias{get_len_data_risk}
\title{Format \code{survdat} length data for risk analysis}
\usage{
get_len_data_risk(x)
}
\arguments{
\item{x}{A \code{survdat} data frame or tibble, containing data on one or more species.}
}
\value{
A tibble
}
\description{
This function formats \code{survdat} length data for subsequent risk analysis. Data from unique observations of a stock (species, region) are averaged by year. Returns lengths in a wide format.
}
