% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survdat_functions.R
\name{get_len_data_tbl}
\alias{get_len_data_tbl}
\title{Format \code{survdat} length data for summary data table}
\usage{
get_len_data_tbl(x)
}
\arguments{
\item{x}{A \code{survdat} data frame or tibble, containing data on one species.}
}
\value{
A tibble
}
\description{
This function formats \code{survdat} length data for a summary data table. Data from unique observations of a stock (species, region) are averaged by year.
}
