% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehensive_risk_report.R
\name{get_risk}
\alias{get_risk}
\title{Calculate risk for comprehensive risk report}
\usage{
get_risk(data, year_source, value_source, analysis, high, indicator_name)
}
\arguments{
\item{data}{A data frame or tibble, containing data on multiple species.}

\item{year_source}{The name of the column with year information.}

\item{value_source}{The name of the column with the value that should be ranked.}

\item{analysis}{The type of analysis to do. Choose from \code{c("most_recent", "past5", "max_alltime", "past10hist", "diet")}.}

\item{high}{Whether a high value is high risk (\code{"high_risk"}) or low risk (\code{"low_risk"}).}

\item{indicator_name}{The name the indicator should have in the output.}
}
\value{
A tibble
}
\description{
This function calculates risk for the comprehensive risk report by ranking species.
}
