% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_functions.R
\name{get_species_risk}
\alias{get_species_risk}
\title{Rank an indicator within a species}
\usage{
get_species_risk(
  data,
  year_source,
  value_source,
  high,
  indicator_name,
  n_run = 5
)
}
\arguments{
\item{data}{A data frame or tibble of indicator data on a single stock.}

\item{year_source}{The name of the column with year data.}

\item{value_source}{The name of the column with indicator data.}

\item{high}{Whether a high value should be assessed as high risk ("high_risk") or low risk ("low_risk")}

\item{indicator_name}{The name the indicator should be assigned in the output.}

\item{n_run}{How many years to calculate the running mean over. Defaults to 5.}
}
\value{
A tibble
}
\description{
This function ranks all measurements of a single indicator throughout the time series.
Measurements are smoothed with a running mean.
}
