% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survdat_functions.R
\name{get_var_data}
\alias{get_var_data}
\title{Format \code{survdat} data}
\usage{
get_var_data(x, variable)
}
\arguments{
\item{x}{A \code{survdat} data frame or tibble, containing data on one or more species.}

\item{variable}{The \code{survdat} measurement of interest. Must be a column name of the \code{survdat} data set.}
}
\value{
A tibble
}
\description{
This function formats \code{survdat} data for subsequent plotting and analysis. Data from unique observations of a stock (species, region) are averaged by year.
}
