% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_spatial_indicators.R
\name{\%>\%}
\alias{\%>\%}
\title{Create Bottom Temperature Indicator}
\usage{
lhs \%>\% rhs

lhs \%>\% rhs

lhs \%>\% rhs

lhs \%>\% rhs

lhs \%>\% rhs
}
\arguments{
\item{data.in}{Either a character vector of full input file names for a list of spatRasters}

\item{output.files}{character vector of full output file names corresponding to each input file}

\item{shp.file}{string. Shape file you wish to crop each input file to}

\item{var.name}{string. Variable name you wish to extract = 'pp'}

\item{area.names}{character vector. Names of shape file areas you want to summarize.}

\item{statistic}{string. Which statistic to calculate = 'mean'}

\item{agg.time}{character. Time scale to calculate over (days, doy, months, season, or years)}

\item{tz}{string. Time zone to convert. No correction if NA}

\item{touches}{logical. If TRUE, all cells touched by lines or polygons will be masked, not just those on the line render path, or whose center point is within the polygon}

\item{write.out}{logical. If TRUE, will write a netCDF file with output.files. If FALSE will return a list of spatRasters}
}
\value{
Saves R object \code{bottomT}, returns bottom temp indicator in a data frame summarized by timestep for each area.names

Saves R object \code{sst}, returns sea surface temp indicator in a data frame summarized by timestep for each area.names

Saves R object \code{sal}, returns salinity indicator in a data frame summarized by timestep for each area.names

Saves R object \code{chl}, returns chlorophyll-a indicator in a data frame summarized by timestep for each area.names

Saves R object \code{pp}, returns primary production indicator in a data frame summarized by timestep for each area.names
}
\description{
This function generates a bottom temperature indicator from a GLORYS netCDF file.
The function passes data through EDABUtilities::make_2d_summary_ts, which provides summary statistics of 2d gridded data as time series by area.
Converts .nc files to data frame.

This function generates a sea surface temperature indicator from an OISST ERDDAP netCDF file.
The function passes data through EDABUtilities::make_2d_summary_ts, which provides summary statistics of 2d gridded data as time series by area.
Converts .nc files to data frame.

This function generates a salinity indicator from a GLORYS netCDF file.
The function passes data through EDABUtilities::make_2d_summary_ts, which provides summary statistics of 2d gridded data as time series by area.
Converts .nc files to data frame.

This function generates a chlorophyll-a indicator from an OCCCI ERDDAP netCDF file.
The function passes data through EDABUtilities::make_2d_summary_ts, which provides summary statistics of 2d gridded data as time series by area.
Converts .nc files to data frame.

This function generates a primary productivity indicator from an ERDDAP netCDF file.
The function passes data through EDABUtilities::make_2d_summary_ts, which provides summary statistics of 2d gridded data as time series by area.
Converts .nc files to data frame.
}
