% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{guild_info}
\alias{guild_info}
\title{Functional guild information}
\format{
A data frame with 36 rows and 4 variables:
\describe{
\item{\code{Species}}{character Species}
\item{\code{Scientific_name}}{character Scientific name}
\item{\code{Guild}}{character Guild}
\item{\code{size}}{character Size (<40cm = small)}
}
}
\usage{
guild_info
}
\description{
Functional guild information
}
\details{
DETAILS
}
\keyword{datasets}
