% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{guild_risk}
\alias{guild_risk}
\title{Guild risk}
\format{
A data frame with 200 rows and 14 variables:
\describe{
\item{\code{Guild}}{character Guild}
\item{\code{size}}{character Size}
\item{\code{Indicator}}{character Indicator}
\item{\code{category}}{character Indicator category}
\item{\code{guild_risk}}{double Sum of risk of all species in the guild for the indicator}
\item{\code{n_guild_species}}{integer Number of species in the guild}
\item{\code{n_guild_species_measured}}{integer Number of species in the guild with indicator measurements}
\item{\code{avg_guild_risk}}{double Average risk of species in the guild}
\item{\code{rank}}{double Rank of the guild, relative to other guilds (ranked by average guild risk)}
\item{\code{norm_rank}}{double Normalized rank}
\item{\code{total_guild_risk}}{double Total guild risk (summed over all indicators)}
\item{\code{sum_ranks}}{double Guild rank summed over all indicators}
\item{\code{legend_label}}{character Legend label (for plotting)}
\item{\code{label_y}}{double Position of y text (for plotting)}
}
}
\usage{
guild_risk
}
\description{
Guild risk, determined by relative rankings
}
\details{
DETAILS
}
\keyword{datasets}
