% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2.R
\docType{data}
\name{ichthyo}
\alias{ichthyo}
\title{DATASET_TITLE}
\format{
A data frame with 6162875 rows and 19 variables:
\describe{
\item{\code{taxa}}{character COLUMN_DESCRIPTION}
\item{\code{year}}{double COLUMN_DESCRIPTION}
\item{\code{season}}{double COLUMN_DESCRIPTION}
\item{\code{strata.x}}{double COLUMN_DESCRIPTION}
\item{\code{lat}}{double COLUMN_DESCRIPTION}
\item{\code{lon}}{double COLUMN_DESCRIPTION}
\item{\code{area}}{double COLUMN_DESCRIPTION}
\item{\code{mean.abund}}{double COLUMN_DESCRIPTION}
\item{\code{rel.proportion}}{double COLUMN_DESCRIPTION}
\item{\code{season.month}}{integer COLUMN_DESCRIPTION}
\item{\code{sci_name}}{character COLUMN_DESCRIPTION}
\item{\code{common_name}}{character COLUMN_DESCRIPTION}
\item{\code{cc_name}}{character COLUMN_DESCRIPTION}
\item{\code{stock_name}}{character COLUMN_DESCRIPTION}
\item{\code{species_code}}{character COLUMN_DESCRIPTION}
\item{\code{svspp}}{integer COLUMN_DESCRIPTION}
\item{\code{stock_season}}{character COLUMN_DESCRIPTION}
\item{\code{strata.y}}{integer COLUMN_DESCRIPTION}
\item{\code{stock_subarea}}{character COLUMN_DESCRIPTION}
}
}
\usage{
ichthyo
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
