% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailored_report_functions.R
\name{ind_rpt}
\alias{ind_rpt}
\title{Create indicator report card output}
\usage{
ind_rpt(var, metric, data)
}
\arguments{
\item{var}{Which level of \code{Var} to plot.}

\item{metric}{The stock metric to assess. Passed to \code{prep_data}.}

\item{data}{The data to use.}
}
\value{
A tibble
}
\description{
This function creates the report card output for an indicator. Variables are passed from \code{wrap_analysis}
}
