% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indicator_info}
\alias{indicator_info}
\title{Indicator risk key}
\format{
A data frame with 24 rows and 5 variables:
\describe{
\item{\code{Category}}{character Indicator category}
\item{\code{Indicator}}{character Indicator}
\item{\code{Data_source}}{character Data source}
\item{\code{Metric}}{character How the indicator was processed for risk assessment}
\item{\code{Risk}}{character How risk was assigned}
}
}
\usage{
indicator_info
}
\description{
Indicator risk key
}
\details{
DETAILS
}
\keyword{datasets}
