% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{latlong}
\alias{latlong}
\title{Geographic distribution}
\format{
A data frame with 3391 rows and 5 variables:
\describe{
\item{\code{stock_season}}{character Season}
\item{\code{strata}}{integer Stratum}
\item{\code{SVSPP}}{integer Northeast Resource Surveys Species code}
\item{\code{Region}}{character Stock region}
\item{\code{Species}}{character Species common name}
}
}
\source{
\url{https://github.com/NOAA-EDAB/ECSA}
}
\usage{
latlong
}
\description{
Seasonal strata information
}
\details{
DETAILS
}
\keyword{datasets}
