% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survdat_functions.R
\name{len_tbl_data}
\alias{len_tbl_data}
\title{Summarise \code{survdat} length data}
\usage{
len_tbl_data(x)
}
\arguments{
\item{x}{A data frame or tibble, containing data on one species. The output of \code{get_len_data_tbl}.}
}
\value{
A \code{DT::datatable}
}
\description{
This function summarises \code{survdat} data, grouping by by region and season. Data must be pre-processed with \code{get_len_data_tbl}.
}
