% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_html_table}
\alias{make_html_table}
\title{Wrapper function for \code{DT::datatable}}
\usage{
make_html_table(x, col_names = colnames(x), type = "html")
}
\arguments{
\item{x}{A data frame or tibble}

\item{col_names}{The column names to display on the output. Defaults to the column names of the data frame.}
}
\value{
An html table
}
\description{
This function is a wrapper for \code{DT::datatable}.
}
