% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailored_report_functions.R
\name{make_ind_rpt}
\alias{make_ind_rpt}
\title{Create indicator report card}
\usage{
make_ind_rpt(data)
}
\arguments{
\item{data}{A summary of indicator(s) of interest.}
}
\value{
A flextable
}
\description{
This function creates an indicator report card for select indicators.
}
