% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailored_report_functions.R
\name{make_time_rpt}
\alias{make_time_rpt}
\title{Create time report card}
\usage{
make_time_rpt(data)
}
\arguments{
\item{data}{A time series of indicator(s) of interest.}
}
\value{
A flextable
}
\description{
This function creates a time series report card for selected indicators.
}
