% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_functions.R
\name{map_strata}
\alias{map_strata}
\title{Map strata}
\usage{
map_strata(common_name, stock_season, strata)
}
\arguments{
\item{stock_season}{A vector of season names (fall, winter, and/or spring)}

\item{save_plot}{}
}
\value{
a ggplot2 object
}
\description{
Copied from \code{NOAA-EDAB/ECSA} but with shapefile input modified.
}
