% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_functions.R
\name{map_strata_ecsa}
\alias{map_strata_ecsa}
\title{Create a map of seasonal species distribution}
\usage{
map_strata_ecsa(data, species_name)
}
\arguments{
\item{data}{A data frame or tibble, containing data on one species. Must contain season and strata information.}

\item{species_name}{The common name of the species}
}
\value{
A ggplot
}
\description{
This function creates a map of species distribution. Built on the \code{ecodata} function \code{map_strata}.
}
