% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_functions.R
\name{missing_na}
\alias{missing_na}
\title{Replace risk rating of 0.5 with NA}
\usage{
missing_na(data)
}
\arguments{
\item{data}{A data frame or tibble of risk data.}
}
\value{
A data frame or tibble
}
\description{
This function replace a risk rating of 0.5 that was assigned due to missing data with "NA".
A helper function.
}
