% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_netcdf_raster.R
\name{nc_to_raster}
\alias{nc_to_raster}
\title{Convert netcdf to raster}
\usage{
nc_to_raster(
  nc,
  varname,
  extent = c(0, 360, -90, 90),
  crop = raster::extent(280, 300, 30, 50),
  show_images = FALSE
)
}
\arguments{
\item{nc}{The nc file path}

\item{varname}{The name of the variable to convert to a raster}

\item{extent}{The latitude and longitude range the data should have, of the form c(xmin, xmax, ymin, ymax). Defaults to \code{c(0, 360, -90, 90)}}

\item{crop}{An extent object to use to crop data. Defaults to \code{raster::extent(280, 300, 30, 50)} (Northeast US)}

\item{show_images}{Boolean. Whether to display images of the data. Useful to check if cropping is occurring correctly.}
}
\value{
A raster brick
}
\description{
This function converts a netcdf object to a raster object
}
