% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nrcc}
\alias{nrcc}
\title{NRCC prioritization}
\format{
A data frame with 43 rows and 7 variables:
\describe{
\item{\code{Species}}{character Species common name}
\item{\code{Region}}{character Stock region}
\item{\code{Indicator}}{character Indicator}
\item{\code{Year}}{character Year when indicator was assessed, or descriptive substitute}
\item{\code{Value}}{double Indicator value}
\item{\code{rank}}{double Species rank by indicator value}
\item{\code{norm_rank}}{double Normalized rank}
}
}
\source{
NRCC prioritization data
}
\usage{
nrcc
}
\description{
Putative NRCC stock prioritization scores
}
\details{
DETAILS
}
\keyword{datasets}
