% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life_history_functions.R
\name{plot_age_density}
\alias{plot_age_density}
\title{Plot \code{survdat} age density data}
\usage{
plot_age_density(data)
}
\arguments{
\item{data}{A \code{survdat} data frame or tibble, containing data on one species.}
}
\value{
A ggplot
}
\description{
This function plots \code{survdat} age density data
}
