% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life_history_functions.R
\name{plot_age_diversity}
\alias{plot_age_diversity}
\title{Plot \code{survdat} age diversity data}
\usage{
plot_age_diversity(data, species)
}
\arguments{
\item{data}{A \code{survdat} data frame or tibble, containing data on one species.}
}
\value{
A ggplot
}
\description{
This function plots \code{survdat} age diversity data
}
