% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessmentdata_functions.R
\name{plot_asmt}
\alias{plot_asmt}
\title{Plot \code{assessmentdata::stockAssessmentData} data}
\usage{
plot_asmt(x, metric, ytitle = "", lin = lines, col = colors)
}
\arguments{
\item{x}{A data frame or tibble, containing data on one species. The output of \code{assessmentdata::stockAssessmentData}.}

\item{metric}{The metric to plot (Recruitment, Abundance, Catch, or F)}

\item{ytitle}{The title of the y-axis. Defaults to "".}

\item{lin}{The linetypes to use for each region (a named vector)}

\item{col}{The colors to use for each region (a named vector)}
}
\value{
A ggplot
}
\description{
This function plots \code{assessmentdata::stockAssessmentData} data.
}
