% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{plot_climate_vulnerability}
\alias{plot_climate_vulnerability}
\title{Plot climate vulnerability ratings}
\usage{
plot_climate_vulnerability(data)
}
\arguments{
\item{x}{A data frame or tibble of climate vulnerability ratings from Hare et al. 2016, containing data on one species.}
}
\value{
A ggplot
}
\description{
This function plots climate vulnerability ratings from Hare et al. 2016.
}
