% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landings_functions.R
\name{plot_com}
\alias{plot_com}
\title{Plot commercial landings data}
\usage{
plot_com(data)
}
\arguments{
\item{data}{Commercial landings data for a single species. Subsetted from FOSS.}
}
\value{
A ggplot
}
\description{
This function plots commercial landings data. Does not contain fine-scale region information.
}
