% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landings_functions.R
\name{plot_com_money}
\alias{plot_com_money}
\title{Plot commercial revenue data}
\usage{
plot_com_money(data)
}
\arguments{
\item{data}{Commercial revenue data for a single species. Subsetted from FOSS.}
}
\value{
A ggplot
}
\description{
This function plots commercial revenue data. Does not contain fine-scale region information.
}
