% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition_functions.R
\name{plot_cond}
\alias{plot_cond}
\title{Plot species condition (weight/length^3)}
\usage{
plot_cond(x)
}
\arguments{
\item{x}{A data frame or tibble, containing data on one species. Data from \code{allfh}.}
}
\value{
A ggplot
}
\description{
This function creates a plot of species condition (weight/length^3).
}
