% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life_history_functions.R
\name{plot_depth}
\alias{plot_depth}
\title{Plot \code{survdat} depth data}
\usage{
plot_depth(data, species)
}
\arguments{
\item{data}{A \code{survdat} data frame or tibble, containing data on one species.}
}
\value{
A ggplot
}
\description{
This function plots \code{survdat} depth data
}
