% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{plot_ecodata}
\alias{plot_ecodata}
\title{Plot stock-level data from \code{ecodata}}
\usage{
plot_ecodata(data, ylabel = "")
}
\arguments{
\item{data}{\code{ecodata} data aggregated at the stock level, filtered to the stock of interest}

\item{ylabel}{Optional. Label for y-axis. Defaults to blank.}
}
\value{
A ggplot
}
\description{
This is a generic function to plot stock-level data from \code{ecodata}
}
