% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survdat_functions.R
\name{plot_len}
\alias{plot_len}
\title{Plot minimum, mean, and maximum \code{survdat} length data}
\usage{
plot_len(x)
}
\arguments{
\item{x}{A data frame or tibble, containing data on one species. The output of \code{get_len_data}.}
}
\value{
A ggplot
}
\description{
This function plots \code{survdat} length data faceted by region. Data must be pre-processed with \code{get_len_data}.
}
