% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survdat_functions.R
\name{plot_len_hist}
\alias{plot_len_hist}
\title{Plot a density curve of \code{survdat} length data}
\usage{
plot_len_hist(x)
}
\arguments{
\item{x}{A data frame or tibble, containing data on one species. A subset of a \code{survdat} pull.}
}
\value{
A ggplot
}
\description{
This function plots \code{survdat} length data faceted by region.
}
