% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition_functions.R
\name{plot_lw}
\alias{plot_lw}
\title{Plot a length-weight curve}
\usage{
plot_lw(x)
}
\arguments{
\item{x}{A data frame or tibble, containing data on one species. Data from \code{allfh}.}
}
\value{
A ggplot
}
\description{
This function creates a length-weight curve.
}
