% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessmentdata_functions.R
\name{plot_msy}
\alias{plot_msy}
\title{Plot B/Bmsy or F/Fmsy from \code{assessmentdata::stockAssessmentSummary}}
\usage{
plot_msy(x, ytitle = "", lin = lines, col = colors, type)
}
\arguments{
\item{x}{A data frame or tibble, containing data on one species. The output of \code{assessmentdata::stockAssessmentSummary}.}

\item{ytitle}{The title of the y-axis. Defaults to "".}

\item{lin}{The linetypes to use for each region (a named vector)}

\item{col}{The colors to use for each region (a named vector)}

\item{type}{"b" for B/Bmsy, "f" for F/Fmsy}
}
\value{
A ggplot
}
\description{
This function plots B/Bmsy or F/Fmsy from \code{assessmentdata::stockAssessmentSummary}. You will get better results from a saved data pull. The most recent version of \code{assessmentdata} does not contain multiple years of msy data.
}
