% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landings_functions.R
\name{plot_prop_catch}
\alias{plot_prop_catch}
\title{Plot the proportion of commercial and recreational landings}
\usage{
plot_prop_catch(rec, com)
}
\arguments{
\item{rec}{Recreational landings data for a single species. Subsetted from MRIP.}

\item{com}{Commercial landings data for a single species. Subsetted from FOSS.}
}
\value{
A ggplot
}
\description{
This function plots the proportion of commercial and recreational landings for all years with existing data. Does not contain fine-scale region information.
}
