% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landings_functions.R
\name{plot_rec}
\alias{plot_rec}
\title{Plot recreational data}
\usage{
plot_rec(data, var, title)
}
\arguments{
\item{data}{Recreational data for a single species. Subsetted from MRIP}

\item{var}{What data to plot. One of c("total_catch2", "discards", "prop", "landings", "landings_num", "prop_land")}

\item{title}{The title for the graph}
}
\value{
A ggplot
}
\description{
This function plots recreational data.
}
