% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landings_functions.R
\name{plot_rec_catch}
\alias{plot_rec_catch}
\title{Plot recreational landings data}
\usage{
plot_rec_catch(data)
}
\arguments{
\item{data}{Recreational landings data for a single species. Subsetted from MRIP}
}
\value{
A ggplot
}
\description{
This function plots commercial recreational data. Does not contain fine-scale region information.
}
