% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition_functions.R
\name{plot_relw}
\alias{plot_relw}
\title{Plot species relative weight}
\usage{
plot_relw(x)
}
\arguments{
\item{x}{A data frame or tibble, containing data on one species. Data from \code{allfh}.}
}
\value{
A ggplot
}
\description{
This function creates a plot of species relative weight.
}
