% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_functions.R
\name{plot_risk_by_stock}
\alias{plot_risk_by_stock}
\title{Plot within-species risk over time}
\usage{
plot_risk_by_stock(data, indicator, include_legend)
}
\arguments{
\item{data}{A data frame or tibble of data. The output of \code{get_species_risk}.}

\item{include_legend}{Whether the plot should have a legend. If blank, legend will be included. If "no", legend will not be included.}

\item{indicator_name}{The name of the indicator to plot.}
}
\value{
A ggplot or multiple ggplots
}
\description{
This function plots within-species risk over time.
Can probably be combined with/deprecated in favor of \code{plot_risk_by_year}.
}
