% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{plot_swept}
\alias{plot_swept}
\title{Plot swept area estimates from \code{survdat}}
\usage{
plot_swept(x, var)
}
\arguments{
\item{x}{A data frame or tibble, containing data on one species. The swept area estimate from \code{survdat}.}

\item{var}{The variable to plot ("abundance" or "biomass")}
}
\value{
A ggplot
}
\description{
This function plots \verb{swept area estimates from }survdat` data.
}
