% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survdat_functions.R
\name{plot_variable}
\alias{plot_variable}
\title{Plot \code{survdat} data}
\usage{
plot_variable(x, ytitle = "")
}
\arguments{
\item{x}{A data frame or tibble, containing data on one species. The output of \code{get_var_data}.}

\item{ytitle}{The title of the y-axis. Defaults to "".}
}
\value{
A ggplot
}
\description{
This function plots \code{survdat} data faceted by region. Data must be pre-processed with \code{get_var_data}.
}
