% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailored_report_functions.R
\name{prep_data}
\alias{prep_data}
\title{Prepare data for plotting with \code{plot_corr_only}}
\usage{
prep_data(data, metric = "Recruitment", pattern = NULL, remove = NULL)
}
\arguments{
\item{data}{A data frame containing stock and indicator time series, or a file path to a .csv containg the data. Data from a spreadsheet outputted by a \code{NEespShiny} or \code{NEesp} regression report.}

\item{metric}{The stock assessment metric to assess - c("Recruitmetn", "Abundance", "Catch", "Fmort")}

\item{pattern}{Optional. A pattern to detect in the \code{Var} row, for example if you do not want to plot all levels of \code{Var}. Can be a vector.}

\item{remove}{Optional. If the pattern should be removed (\code{TRUE}) or retained (\code{FALSE}). Can be a vector.}
}
\value{
A tibble
}
\description{
This function prepares data for plotting with \code{plot_corr_only}.
}
