% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailored_report_functions.R
\name{prep_si_data}
\alias{prep_si_data}
\title{Prepare data for plotting with \code{plot_corr_only}, within shiny app}
\usage{
prep_si_data(file_path, metric = "Recruitment", var)
}
\arguments{
\item{file_path}{The file path to the data. Data from a spreadsheet outputted by a \code{NEespShiny} or \code{NEesp} regression report.}

\item{metric}{The stock assessment metric to assess - c("Recruitment", "Abundance", "Catch", "Fmort")}

\item{var}{Which level of the \code{Var} column to plot.}
}
\value{
A tibble
}
\description{
This function prepares data for plotting with \code{plot_corr_only}.
}
