% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prey_palette}
\alias{prey_palette}
\title{Diet data color palette}
\format{
A data frame with 19 rows and 2 variables:
\describe{
\item{\code{prey_id}}{character Prey ID}
\item{\code{color}}{character Color (hex code)}
}
}
\usage{
prey_palette
}
\description{
A color palette to use with diet data
}
\details{
Created from \code{nmfspalette} package
}
\keyword{datasets}
