% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landings_functions.R
\name{prop_catch_data}
\alias{prop_catch_data}
\title{Returns proportional commercial and recreational landings data}
\usage{
prop_catch_data(rec, com)
}
\arguments{
\item{rec}{Recreational landings data for a single species. Subsetted from MRIP.}

\item{com}{Commercial landings data for a single species. Subsetted from FOSS.}
}
\value{
A tibble
}
\description{
This function joins for formats commercial and recreational landings data for all years with existing data. Does not contain fine-scale region information.
}
