% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_functions.R
\name{pseudoR2}
\alias{pseudoR2}
\title{Calculate a pseudo-R2 for Poisson regressions}
\usage{
pseudoR2(model, data, respv_colname)
}
\arguments{
\item{model}{A Poisson GLM model fit to the data}

\item{data}{The data used to fit the model}

\item{respv_colname}{The response variable column name}
}
\value{
A pseudo-R2 value
}
\description{
This function calculates a pseudo-R2 for Poisson regressions.
}
