% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2.R
\docType{data}
\name{rec_catch_small}
\alias{rec_catch_small}
\title{DATASET_TITLE}
\format{
A data frame with 63710 rows and 7 variables:
\describe{
\item{\code{year}}{integer COLUMN_DESCRIPTION}
\item{\code{sub_reg_f}}{character COLUMN_DESCRIPTION}
\item{\code{st_f}}{character COLUMN_DESCRIPTION}
\item{\code{mode_fx_f}}{character COLUMN_DESCRIPTION}
\item{\code{tot_cat}}{integer COLUMN_DESCRIPTION}
\item{\code{lbs_ab1}}{double COLUMN_DESCRIPTION}
\item{\code{Species}}{character COLUMN_DESCRIPTION}
}
}
\usage{
rec_catch_small
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
