% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landings_functions.R
\name{rec_data_prep}
\alias{rec_data_prep}
\title{Prepare recreational data for plotting}
\usage{
rec_data_prep(data)
}
\arguments{
\item{data}{Recreational data for a single species. Subsetted from MRIP}
}
\value{
A tibble
}
\description{
This function does basic data manipulation on recreational data.
}
