% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2.R
\docType{data}
\name{rec_effort}
\alias{rec_effort}
\title{DATASET_TITLE}
\format{
A data frame with 14478 rows and 8 variables:
\describe{
\item{\code{Estimate.Status}}{integer COLUMN_DESCRIPTION}
\item{\code{Year}}{integer COLUMN_DESCRIPTION}
\item{\code{Wave}}{integer COLUMN_DESCRIPTION}
\item{\code{State}}{integer COLUMN_DESCRIPTION}
\item{\code{Fishing.Mode}}{integer COLUMN_DESCRIPTION}
\item{\code{Fishing.Area}}{integer COLUMN_DESCRIPTION}
\item{\code{Angler.Trips}}{integer COLUMN_DESCRIPTION}
\item{\code{PSE}}{integer COLUMN_DESCRIPTION}
}
}
\usage{
rec_effort
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
