% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rec_palette}
\alias{rec_palette}
\title{Recreational landings data color palette}
\format{
A data frame with 5 rows and 2 variables:
\describe{
\item{\code{rec_mode}}{character Fishing mode}
\item{\code{color}}{character Color (hex code)}
}
}
\usage{
rec_palette
}
\description{
A color palette to use with recreational landings dat
}
\details{
Created from \code{nmfspalette} package
}
\keyword{datasets}
