% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regression_species_regions}
\alias{regression_species_regions}
\title{Species, stock region, and EPU combinations}
\format{
A data frame with 61 rows and 3 variables:
\describe{
\item{\code{Species}}{character Species common name}
\item{\code{Region}}{character Stock region}
\item{\code{EPU}}{character Environmental Protection Unit}
}
}
\usage{
regression_species_regions
}
\description{
Key to link species, stock regions, and EPUs in all plausible combinations.
}
\details{
DETAILS
}
\keyword{datasets}
