% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_regression_reports.R
\name{render_all_reg}
\alias{render_all_reg}
\title{Renders all regression ESP reports}
\usage{
render_all_reg(x = "package")
}
\arguments{
\item{x}{The folder with the bookdown template. Defaults to "package", which calls the template files saved in the package.}
}
\value{
Multiple bookdown reports (html)
}
\description{
This function renders all regression ESP reports.
}
