% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehensive_risk_report.R
\name{render_comp_risk}
\alias{render_comp_risk}
\title{Render the comprehensive risk report}
\usage{
render_comp_risk(folder)
}
\arguments{
\item{folder}{The folder where the output file should be saved.}
}
\value{
An html
}
\description{
This function the comprehensive risk report
}
