% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_indicator_reports.R
\name{render_ind_report}
\alias{render_ind_report}
\title{Renders an indicator ESP report}
\usage{
render_ind_report(x, trouble = FALSE)
}
\arguments{
\item{x}{The common name of the species}

\item{trouble}{Whether or not to display verbose output. Defaults to FALSE.}
}
\value{
A bookdown report (html) (saved in a folder called \code{action_reports} in the root directory)
}
\description{
This function renders an indicator ESP report.
}
