% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_functions.R
\name{render_indicator}
\alias{render_indicator}
\title{Renders an RMarkdown section}
\usage{
render_indicator(test)
}
\arguments{
\item{test}{A data table from \code{ecodata} that has been pre-processed to a standardized format.}
}
\value{
An RMarkdown section
}
\description{
This function renders an RMarkdown section, including plots and data tables, for stock data and an \code{ecodata} indicator. Designed for use in an RMarkdown document. Stock data must be pre-defined.
}
