% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_regression_reports.R
\name{render_reg_report}
\alias{render_reg_report}
\title{Renders a regression ESP report}
\usage{
render_reg_report(
  stock_var,
  epus_var,
  region_var,
  remove_var = FALSE,
  lag_var = 0,
  parent_folder,
  input = "package",
  trouble = FALSE,
  save_var = TRUE
)
}
\arguments{
\item{stock_var}{The name of the species}

\item{epus_var}{The name of the EPU ("MAB", "GB", or "GOM")}

\item{region_var}{The name of the region the species is found in (must match an option in \code{assessmentdata::stockAssessmentData})}

\item{remove_var}{Whether or not to remove the most recent 10 years of data. Defaults to FALSE.}

\item{lag_var}{How many years to lag the correlations by. Defaults to 0.}

\item{parent_folder}{The name of the folder to put the output in. If it does not exist, it will be created.}

\item{input}{The folder with the bookdown template. Defaults to "package", which calls the template files saved in the package.}

\item{trouble}{Whether or not to display verbose output. Defaults to FALSE.}

\item{save_var}{Whether or not to save the data used in report creation. Defaults to TRUE.}
}
\value{
A bookdown report (html) (saved in a folder called \code{Regressions} in the root directory)
}
\description{
This function renders a regression ESP report.
}
