% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{risk}
\alias{risk}
\title{Overall risk assessment}
\format{
A data frame with 1375 rows and 15 variables:
\describe{
\item{\code{Species}}{character Species common name}
\item{\code{Region}}{character Stock region}
\item{\code{Indicator}}{character Indicator}
\item{\code{category}}{character Indicator category}
\item{\code{Year}}{character Year, time-based description, or other qualitative label if year does not apply}
\item{\code{Value}}{double Value}
\item{\code{rank}}{character Stock's indicator rank relative to all other stocks}
\item{\code{n_stocks_per_indicator}}{integer Number of stocks compared for the indicator}
\item{\code{n_stocks_per_region}}{integer Number of stocks in the stock region}
\item{\code{norm_rank}}{double Normalized rank}
\item{\code{total_risk}}{double Stock's total risk (sum of normalized ranks for all indicators)}
\item{\code{overall_rank}}{integer Stock rank of total risk, compared to all other stocks}
\item{\code{overall_stocks}}{integer Total number of stocks}
\item{\code{stock}}{character Species and region}
\item{\code{label}}{character Label (for plotting)}
}
}
\usage{
risk
}
\description{
Species-level risk assessment with a variety of indicator analysis methods
}
\details{
DETAILS
}
\keyword{datasets}
