% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{risk_species}
\alias{risk_species}
\title{Risk assessment within stock}
\format{
A data frame with 50880 rows and 12 variables:
\describe{
\item{\code{Species}}{character Species common name}
\item{\code{Region}}{character Stock region}
\item{\code{Indicator}}{character Indicator}
\item{\code{category}}{character Indicator category}
\item{\code{Year}}{character Window of years}
\item{\code{Value}}{double Value}
\item{\code{rank}}{double Stock's indicator rank for the given year relative to all other years of the same stock}
\item{\code{n_years_per_indicator}}{double Number of years ranked for the stock}
\item{\code{n_stocks_per_region}}{integer NUmber of stocks in the stock region}
\item{\code{norm_rank}}{double Normalized rank}
\item{\code{total_risk}}{double Stock's total risk (sum of normalized ranks for all indicators)}
\item{\code{overall_rank}}{integer Stock's rank of total risk, compared to all other stocks}
}
}
\usage{
risk_species
}
\description{
Risk assessment: years ranked within a single stock
}
\details{
DETAILS
}
\keyword{datasets}
