% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{risk_year_hist}
\alias{risk_year_hist}
\title{Risk assessment of change compared to historical value}
\format{
A data frame with 41658 rows and 15 variables:
\describe{
\item{\code{Species}}{character Species common name}
\item{\code{Region}}{character Stock region}
\item{\code{Indicator}}{character Indicator}
\item{\code{category}}{character Indicator category}
\item{\code{Year}}{character Window of years}
\item{\code{Value}}{double Value (change compared to historical)}
\item{\code{rank}}{double Stock's indicator rank for the given year relative to all other stocks in the same year}
\item{\code{n_stocks_per_indicator}}{double Number of stocks ranked for the indicator}
\item{\code{n_stocks_per_region}}{integer NUmber of stocks in the stock region}
\item{\code{norm_rank}}{double Normalized rank}
\item{\code{total_risk}}{double  Stock's total risk (sum of normalized ranks for all indicators)}
\item{\code{overall_rank}}{integer Stock's rank of total risk, compared to all other stocks}
\item{\code{overall_stocks}}{integer Total number of stocks}
\item{\code{stock}}{character Species and region}
\item{\code{label}}{character Label (for plotting)}
}
}
\usage{
risk_year_hist
}
\description{
Risk assessment: present year value compared to historical value compared among stocks
}
\details{
DETAILS
}
\keyword{datasets}
