% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_data}
\alias{save_data}
\title{Save a data set}
\usage{
save_data(x)
}
\arguments{
\item{x}{A data table or tibble}
}
\value{
A .csv (small files), or a .RDS (large files)
}
\description{
This function saves a data set if it has nrow > 0. Data is saved in a folder called \code{data} in the working directory.
}
