% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shape}
\alias{shape}
\title{Northeast strata shapefile}
\format{
A data frame with 185 rows and 8 variables:
\describe{
\item{\code{FINSTR_ID}}{integer COLUMN_DESCRIPTION}
\item{\code{STRATA}}{integer Strata ID}
\item{\code{A2}}{integer COLUMN_DESCRIPTION}
\item{\code{STR2}}{integer COLUMN_DESCRIPTION}
\item{\code{SET_}}{integer COLUMN_DESCRIPTION}
\item{\code{STRATUMA}}{character COLUMN_DESCRIPTION}
\item{\code{STR3}}{integer COLUMN_DESCRIPTION}
\item{\code{geometry}}{list Geometry}
}
}
\usage{
shape
}
\description{
Shapefile containing geometries for NE stock strata
}
\details{
DETAILS
}
\keyword{datasets}
