% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2.R
\docType{data}
\name{sp_group}
\alias{sp_group}
\title{DATASET_TITLE}
\format{
A data frame with 1023 rows and 14 variables:
\describe{
\item{\code{COMNAME}}{character COLUMN_DESCRIPTION}
\item{\code{SVSPP}}{integer COLUMN_DESCRIPTION}
\item{\code{ITISSPP}}{integer COLUMN_DESCRIPTION}
\item{\code{NESPP3}}{integer COLUMN_DESCRIPTION}
\item{\code{SCINAME}}{character COLUMN_DESCRIPTION}
\item{\code{Fed_Managed}}{character COLUMN_DESCRIPTION}
\item{\code{SOE_17}}{character COLUMN_DESCRIPTION}
\item{\code{EMAX}}{character COLUMN_DESCRIPTION}
\item{\code{RPATH}}{character COLUMN_DESCRIPTION}
\item{\code{SOE_18}}{character COLUMN_DESCRIPTION}
\item{\code{SizeCat}}{character COLUMN_DESCRIPTION}
\item{\code{Min_size}}{integer COLUMN_DESCRIPTION}
\item{\code{Garrison_Link}}{character COLUMN_DESCRIPTION}
\item{\code{NEIEA}}{character COLUMN_DESCRIPTION}
}
}
\usage{
sp_group
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
