% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species_guilds}
\alias{species_guilds}
\title{Species guild key}
\format{
A data frame with 37 rows and 3 variables:
\describe{
\item{\code{Species}}{character Species common name}
\item{\code{Scientific_name}}{character Species scientific name}
\item{\code{Guild}}{character Guild}
}
}
\usage{
species_guilds
}
\description{
Species guild key
}
\details{
Can probably be deprecated in favor of \code{guild_info}
}
\keyword{datasets}
