% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species_key}
\alias{species_key}
\title{Species SVSPP key}
\format{
A data frame with 37 rows and 2 variables:
\describe{
\item{\code{SVSPP}}{integer Northeast Resource Surveys Species code}
\item{\code{Species}}{character Species common name}
}
}
\usage{
species_key
}
\description{
A key to connect species common names and SVSPP (Northeast Resource Surveys Species code)
}
\details{
DETAILS
}
\keyword{datasets}
