% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessmentdata_functions.R
\name{status}
\alias{status}
\title{Assess the status of B/Bmsy or F/Fmsy from \code{assessmentdata::stockAssessmentSummary}}
\usage{
status(data, regions, metric)
}
\arguments{
\item{data}{A data frame or tibble, containing data on one species. The output of \code{assessmentdata::stockAssessmentSummary}.}

\item{regions}{The region to assess.}

\item{metric}{"bbmsy" for B/Bmsy, "ffmsy" for F/Fmsy}
}
\value{
A character
}
\description{
This function assesses the status of the most recent B/Bmsy or F/Fmsy from \code{assessmentdata::stockAssessmentSummary}.
}
