% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{swept}
\alias{swept}
\title{Swept area estimates}
\format{
A data frame with 4009 rows and 17 variables:
\describe{
\item{\code{YEAR}}{integer Year}
\item{\code{SVSPP}}{integer Northeast Resource Surveys Species code}
\item{\code{N}}{integer COLUMN_DESCRIPTION}
\item{\code{strat.biomass}}{double Stratified mean biomass}
\item{\code{biomass.var}}{double Stratified mean biomass variance}
\item{\code{biomass.SE}}{double Stratified mean biomass standard error}
\item{\code{strat.abund}}{double Stratified mean abundance}
\item{\code{abund.var}}{double Stratified mean abundance variance}
\item{\code{abund.SE}}{double Stratified mean abundance standard error}
\item{\code{tot.biomass}}{double Swept area biomass}
\item{\code{tot.bio.var}}{double Biomass variance}
\item{\code{tot.bio.SE}}{double Biomass standard error}
\item{\code{tot.abundance}}{double Swept area abundance}
\item{\code{tot.abund.var}}{double Abundance variance}
\item{\code{tot.abund.SE}}{double Abundance standard error}
\item{\code{Species}}{character Species common name}
\item{\code{Season}}{character Season}
}
}
\usage{
swept
}
\description{
Swept area estimates
}
\details{
See \code{survdat::swept_area}
}
\keyword{datasets}
