% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailored_report_functions.R
\name{time_rpt}
\alias{time_rpt}
\title{Create an indicator report card rating for one indicator}
\usage{
time_rpt(data, out_name = "unnamed", min_year = 2016)
}
\arguments{
\item{data}{A data frame containing stock and indicator time series. Data from a spreadsheet outputted by a \code{NEespShiny} or \code{NEesp} regression report. Must pre-process with \code{NEesp::prep_data}, or do your own pre-processing.}

\item{out_name}{The name that the indicator column should have in the output.}

\item{min_year}{The minimum year to consider for the recent time-series average. Defaults to 2016.}
}
\value{
A tibble
}
\description{
This function creates an indicator report card. The report card assesses ecosystem/socioeconomic favorability for the stock, NOT overall ecosystem/socioeconomic favorability.
}
