% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailored_report_functions.R
\name{wrap_analysis}
\alias{wrap_analysis}
\title{Wrapper function to prep data, plot data, and create report card output}
\usage{
wrap_analysis(
  file_path,
  metric = "Recruitment",
  var,
  lag = 0,
  min_year = 2016,
  species = "species",
  mode = "download"
)
}
\arguments{
\item{file_path}{The file path to a spreadsheet outputted by a \code{NEespShiny} or \code{NEesp} regression report.}

\item{metric}{The stock metric to assess. Passed to \code{prep_data}.}

\item{var}{Which level of \code{Var} to plot.}

\item{lag}{Passed to \code{plot_corr_only}. The number of years by which the stock-indicator correlation was lagged. Required to correct the stock time series. Defaults to 0.}

\item{min_year}{Passed to \code{time_rpt}. The minimum year to consider for the recent time-series average. Defaults to 2016.}

\item{species}{Passed to \code{plot_corr_only}. The species name to add to plots. Defaults to "species".}

\item{mode}{If set to "shiny", plots will be displayed but no other functionality will be triggered (ex, saving figures or showing a report card). Also passed to \code{plot_corr_only}.}
}
\value{
A tibble
}
\description{
This is a wrapper function of \code{prep_data}, \code{plot_corr_only}, and \code{time_rpt}
}
