% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2.R
\docType{data}
\name{zoop_stock}
\alias{zoop_stock}
\title{DATASET_TITLE}
\format{
A data frame with 128 rows and 35 variables:
\describe{
\item{\code{ctyp_100m3}}{double COLUMN_DESCRIPTION}
\item{\code{calfin_100m3}}{double COLUMN_DESCRIPTION}
\item{\code{pseudo_100m3}}{double COLUMN_DESCRIPTION}
\item{\code{penilia_100m3}}{double COLUMN_DESCRIPTION}
\item{\code{tlong_100m3}}{double COLUMN_DESCRIPTION}
\item{\code{cham_100m3}}{double COLUMN_DESCRIPTION}
\item{\code{echino_100m3}}{double COLUMN_DESCRIPTION}
\item{\code{larvaceans_100m3}}{double COLUMN_DESCRIPTION}
\item{\code{para_100m3}}{double COLUMN_DESCRIPTION}
\item{\code{gas_100m3}}{double COLUMN_DESCRIPTION}
\item{\code{acarspp_100m3}}{double COLUMN_DESCRIPTION}
\item{\code{mlucens_100m3}}{double COLUMN_DESCRIPTION}
\item{\code{evadnespp_100m3}}{double COLUMN_DESCRIPTION}
\item{\code{salps_100m3}}{double COLUMN_DESCRIPTION}
\item{\code{oithspp_100m3}}{double COLUMN_DESCRIPTION}
\item{\code{chaeto_100m3}}{double COLUMN_DESCRIPTION}
\item{\code{hyper_100m3}}{double COLUMN_DESCRIPTION}
\item{\code{calminor_100m3}}{double COLUMN_DESCRIPTION}
\item{\code{clauso}}{double COLUMN_DESCRIPTION}
\item{\code{dec_100m3}}{double COLUMN_DESCRIPTION}
\item{\code{euph_100m3}}{double COLUMN_DESCRIPTION}
\item{\code{prot_100m3}}{double COLUMN_DESCRIPTION}
\item{\code{poly_100m3}}{double COLUMN_DESCRIPTION}
\item{\code{fish_100m3}}{double COLUMN_DESCRIPTION}
\item{\code{oncaea_100m3}}{double COLUMN_DESCRIPTION}
\item{\code{cory_100m3}}{double COLUMN_DESCRIPTION}
\item{\code{siph_100m3}}{double COLUMN_DESCRIPTION}
\item{\code{coel_100m3}}{double COLUMN_DESCRIPTION}
\item{\code{euph1_100m3}}{double COLUMN_DESCRIPTION}
\item{\code{thecos_100m3}}{double COLUMN_DESCRIPTION}
\item{\code{spirspp_100m3}}{double COLUMN_DESCRIPTION}
\item{\code{cnidar_100m3}}{double COLUMN_DESCRIPTION}
\item{\code{Species}}{character COLUMN_DESCRIPTION}
\item{\code{Season}}{character COLUMN_DESCRIPTION}
\item{\code{Time}}{character COLUMN_DESCRIPTION}
}
}
\usage{
zoop_stock
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
