% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{asmt}
\alias{asmt}
\title{Diet and Condition Data}
\format{
A data frame with 1423676 rows and 16 variables:
\describe{
\item{\code{pynam}}{character Prey scientific name}
\item{\code{gencat}}{character Prey general category (shortened)}
\item{\code{gensci}}{character Prey general category (full)}
\item{\code{analcat}}{character Prey analytical category (shortened)}
\item{\code{analsci}}{character Prey analytical category (full)}
\item{\code{pdlen}}{double Predtaor length}
\item{\code{pdwgt}}{integer Predator weight}
\item{\code{stratum}}{integer Stratum}
\item{\code{surftemp}}{double Surface water temperature}
\item{\code{bottemp}}{double Bottom water temperature}
\item{\code{year}}{integer Year}
\item{\code{season}}{integer Season}
\item{\code{pyamtw}}{double Prey weight}
\item{\code{Species}}{character Predator species}
\item{\code{Region}}{character Predator region}
\item{\code{fish_id}}{character Predator ID number}
}

A data frame with 35636 rows and 10 variables:
\describe{
\item{\code{Species}}{character The species' common name}
\item{\code{Region}}{character The stock region}
\item{\code{Year}}{double The year of the value estimate}
\item{\code{Value}}{double The value of the metric}
\item{\code{Metric}}{character The stock assessment metric (catch, F, abundance, or recruitment)}
\item{\code{Description}}{character Details about the metric}
\item{\code{Units}}{character The units of the metric}
\item{\code{AssessmentYear}}{double The year that the assessment was conducted}
\item{\code{Age}}{character The age of the fish (recruitment/abundance only)}
\item{\code{Category}}{character COLUMN_DESCRIPTION}
}
}
\source{
\url{https://github.com/Laurels1/Condition/blob/master/data/allfh.RData}

\url{https://github.com/NOAA-EDAB/assessmentdata}
}
\usage{
asmt
}
\description{
This data was derived from NEFSC bottom trawl data. It contains stomach content information and length-weight information.

This data set contains stock assessment data on several US stocks.
}
\details{
DETAILS
"allfh"

DETAILS
}
\keyword{datasets}
