% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{asmt_sum}
\alias{asmt_sum}
\title{Stock Assessment Summary Data}
\format{
A data frame with 337 rows and 12 variables:
\describe{
\item{\code{FSSI Stock?}}{character Is the stock included in the Fish Stock Sustainability Index?}
\item{\code{Assessment Year}}{integer Assessment year}
\item{\code{Last Data Year}}{integer Last data year used in assessment}
\item{\code{Review Result}}{character Results of assessment review by the Council}
\item{\code{Life History Data}}{integer Biological data rating (before 2019): 0 = no data ; 4 = sufficient data for informed modeling}
\item{\code{Biological Input Data}}{logical Biological data rating (2019 onwards): 0 = no data ; 5 = detailed data with no gaps}
\item{\code{Ecosystem Linkage}}{integer Use of ecosystem linkage data in stock assessment model: 0 = not used ; 5 = model is linked to ecosystem processes}
\item{\code{Composition Input Data}}{integer Availability of size/age data for the stock assessment model: 0 = none ; 5 = complete data with spatial resolution}
\item{\code{F/Fmsy}}{double Fishing pressure relative to the fishing pressure for maximum sustainable yield}
\item{\code{B/Bmsy}}{double Biomass relative to the biomass for maximum sustainable yield}
\item{\code{Species}}{character Species common name}
\item{\code{Region}}{character Stock region}
}
}
\source{
\url{https://github.com/NOAA-EDAB/assessmentdata}
}
\usage{
asmt_sum
}
\description{
This data set contains summary stock assessment metrics on several US stocks.
}
\details{
For more details see \url{https://www.st.nmfs.noaa.gov/stocksmart/StockSMART_DataDictionary.pdf}
}
\keyword{datasets}
