% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bio_survey}
\alias{bio_survey}
\title{Stock Assessment Summary Data}
\format{
A data frame with 337 rows and 12 variables:
\describe{
\item{\code{FSSI Stock?}}{character Is the stock included in the Fish Stock Sustainability Index?}
\item{\code{Assessment Year}}{integer Assessment year}
\item{\code{Last Data Year}}{integer Last data year used in assessment}
\item{\code{Review Result}}{character Results of assessment review by the Council}
\item{\code{Life History Data}}{integer Biological data rating (before 2019): 0 = no data ; 4 = sufficient data for informed modeling}
\item{\code{Biological Input Data}}{logical Biological data rating (2019 onwards): 0 = no data ; 5 = detailed data with no gaps}
\item{\code{Ecosystem Linkage}}{integer Use of ecosystem linkage data in stock assessment model: 0 = not used ; 5 = model is linked to ecosystem processes}
\item{\code{Composition Input Data}}{integer Availability of size/age data for the stock assessment model: 0 = none ; 5 = complete data with spatial resolution}
\item{\code{F/Fmsy}}{double Fishing pressure relative to the fishing pressure for maximum sustainable yield}
\item{\code{B/Bmsy}}{double Biomass relative to the biomass for maximum sustainable yield}
\item{\code{Species}}{character Species common name}
\item{\code{Region}}{character Stock region}
}

A data frame with 1103575 rows and 28 variables:
\describe{
\item{\code{CRUISE6}}{character Cruise ID}
\item{\code{STATION}}{character Station ID}
\item{\code{STRATUM}}{character Stratum}
\item{\code{SVSPP}}{character Northeast Resource Surveys Species code}
\item{\code{CATCHSEX}}{character Species-specific numeric sex code}
\item{\code{LENGTH}}{double Length}
\item{\code{TOW}}{character Tow ID}
\item{\code{SVVESSEL}}{character Vessel ID}
\item{\code{YEAR}}{double Year}
\item{\code{SEASON}}{character Season}
\item{\code{LAT}}{double Latitude}
\item{\code{LON}}{double Longitude}
\item{\code{EST_TOWDATE}}{double Date with estimated time}
\item{\code{DEPTH}}{double Depth}
\item{\code{SURFTEMP}}{double Surface water temperature}
\item{\code{SURFSALIN}}{double Surface salinity}
\item{\code{BOTTEMP}}{double Bottom water temperature}
\item{\code{BOTSALIN}}{double Bottom salinity}
\item{\code{ABUNDANCE}}{double Abundance (number of fish caught on a single tow; redundant for rows with the same CRUISE6, STATION, and TOW)}
\item{\code{BIOMASS}}{double Biomass (biomass caught on a single tow; redundant for rows with the same SVSPP, CRUISE6, STATION, and TOW)}
\item{\code{NUMLEN}}{double Number of fish at the specified length (redundant for rows with the same SVSPP, CRUISE6, STATION, TOW, and LENGTH)}
\item{\code{INDID}}{double Individual fish ID; unknown if actually unique}
\item{\code{INDWT}}{double Individual fish weight}
\item{\code{SEX}}{character Sex}
\item{\code{MATURITY}}{character Maturity}
\item{\code{AGE}}{double Age}
\item{\code{STOM_VOLUME}}{double Stomach volume}
\item{\code{STOM_WGT}}{double Stomach weight}
}
}
\source{
\url{https://github.com/NOAA-EDAB/assessmentdata}
}
\usage{
bio_survey
}
\description{
This data set contains summary stock assessment metrics on several US stocks.

This data is NEFSC bottom trawl data with bio data.
}
\details{
For more details see \url{https://www.st.nmfs.noaa.gov/stocksmart/StockSMART_DataDictionary.pdf}
"asmt_sum"

DETAILS
}
\keyword{datasets}
