% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_functions.R
\name{correlation_data}
\alias{correlation_data}
\title{Produces summary tables of correlations}
\usage{
correlation_data(stock, eco, lag = 0)
}
\arguments{
\item{stock}{Data about a single stock (one species, one region) subsetted from \code{assessmentdata::stockAssessmentData}}

\item{eco}{A data table from \code{ecodata}. May require pre-processing to standardize format.}

\item{lag}{The number of years to lag the correlation by. Defaults to 0.}
}
\value{
A data frame
}
\description{
This function creates summary tables of correlations between stock data and \code{ecodata} data. Designed for use within a RMarkdown document.
}
