% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_functions.R
\name{correlation_summary}
\alias{correlation_summary}
\title{Produces a summary table of correlations}
\usage{
correlation_summary(stock, eco, lag = 0)
}
\arguments{
\item{stock}{Data about a single stock (one species, one region) subsetted from \code{assessmentdata::stockAssessmentData}}

\item{eco}{A data table from \code{ecodata}. May require pre-processing to standardize format.}

\item{lag}{The number of years to lag the correlation by. Defaults to 0.}
}
\value{
A tibble
}
\description{
This function creates summary table of a correlation between stock data and \code{ecodata} data. Suggest to use with multiple \code{ecodata} indicators, and create a master data set by appending results.
}
