% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{get_diet_table}
\alias{get_diet_table}
\title{Create a table of diet data}
\usage{
get_diet_table(data)
}
\arguments{
\item{x}{A data frame or tibble, containing data on one species. Data from \code{allfh}.}
}
\value{
A \code{DT::datatable}
}
\description{
This function creates a table of the proportional composition of a species' diet.
}
