% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_functions.R
\name{map_strata}
\alias{map_strata}
\title{Create a map of seasonal species distribution}
\usage{
map_strata(common_name, strata)
}
\arguments{
\item{common_name}{The common name of the species.}

\item{strata}{A data frame or tibble, containing data on one species. Must contain season and strata information.}
}
\value{
A ggplot
}
\description{
This function creates a map of species distribution. Built from the \code{ecodata} function \code{map_strata}.
}
