% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_functions.R
\name{plot_correlation}
\alias{plot_correlation}
\title{Plots correlations between stock data and indicators}
\usage{
plot_correlation(stock, eco, lag = 0)
}
\arguments{
\item{stock}{Data about a single stock (one species, one region) subsetted from \code{assessmentdata::stockAssessmentData}}

\item{eco}{A data table from \code{ecodata}. May require pre-processing to standardize format.}

\item{lag}{The number of years to lag the correlation by. Defaults to 0.}
}
\value{
A ggplot
}
\description{
This function plots correlations between stock data and \code{ecodata} data.
}
