% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_indicator_reports.R
\name{render_ind_report}
\alias{render_ind_report}
\title{Renders an indicator ESP report}
\usage{
render_ind_report(
  x,
  input = "package",
  params_to_use,
  trouble = FALSE,
  save_data = TRUE,
  out = bookdown::gitbook(split_by = "section", fig_caption = TRUE)
)
}
\arguments{
\item{x}{The common name of the species. \code{NEesp::species_key$Species} shows all available options.}

\item{input}{The folder with the bookdown template (include full file path, e.g. use \code{here::here}). Defaults to "package", which calls the template files saved in the package.}

\item{params_to_use}{A list of parameters to use in the markdown report. Do not set if using \code{input = "package"}.}

\item{trouble}{Whether or not to display verbose output. Defaults to FALSE.}

\item{save_data}{Whether or not to save the data used in report creation. Only relevant when using \code{input = "package"}. Defaults to TRUE.}

\item{out}{The \code{output_format} to be passed to the \code{bookdown::render_book} call. Defaults to \code{bookdown::gitbook(split_by = "section", fig_caption = TRUE)}. For experimental purposes only: most alternative outputs have not been tested.}
}
\value{
A bookdown report (html) (saved in a folder called \code{action_reports} in the root directory)
}
\description{
This function renders an indicator ESP report.
}
