% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_species_names}
\alias{update_species_names}
\title{Change "Goosefish" to "Monkfish" in a designated "Species" column}
\usage{
update_species_names(data, species_col)
}
\arguments{
\item{data}{A data frame or tibble}

\item{species_col}{The name of the column with species names}
}
\value{
A tibble
}
\description{
This function changes any instances of "Goosefish" to "Monkfish" in a designated "Species" column
}
